backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.3
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.10.0
uritemplate==4.1.1
